backend/requirements.txt
# Reto Prosperas - Merged Dependencies
# ====================================
# Combined requirements for both backend and worker services

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AWS SDK
aiobotocore>=2.12.0
boto3>=1.34.0

# Async HTTP
aiohttp>=3.9.0

# Environment & Configuration
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Structured Logging
structlog>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# CloudWatch Observability
watchtower>=3.0.0
